+++ src/mem_arena.c
@@ -1,8 +1,9 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stddef.h>
 #include "../include/mem_arena.h"
 
-MEMARENA* init_arena(int mem_size){
+MEMARENA* init_arena(size_t mem_size){
     MEMARENA* arena = malloc(sizeof(MEMARENA));
     if(arena != NULL){
         arena->base = malloc(mem_size);
@@ -17,7 +18,7 @@ MEMARENA* init_arena(int mem_size){
     return arena;
 }
 
-void* alloc_arena(MEMARENA* arena,int block_size){
+void* alloc_arena(MEMARENA* arena,size_t block_size){
     void* block = NULL;
     if(arena != NULL && arena->base != NULL){
         block = arena->base + arena->offset;
